deployment purpose

+++ backend/src/main/java/com/example/demo/Map/dto/RideData.java
@@ -16,5 +16,5 @@ public class RideData {
     private Long rideId;
     private Integer selectedMarkerIndex1;
     private Integer selectedMarkerIndex2;
-    private Long createdBy;
+    private Integer createdBy;
 }
